requirements.txt
BeautifulSoup==3.2.1
Django==1.6.5
Pillow==2.5.1
PyYAML==3.11
South==0.8.4
Unidecode==0.04.16
argparse==1.2.1
beautifulsoup4==4.3.2
bottlenose==0.6.1
chardet==2.2.1
coverage==3.6
cssselect==0.9.1
dj-database-url==0.3.0
dj-static==0.0.6
django-annoying==0.7.9
django-appconf==0.6
django-common-helpers==0.6.1
django-compressor==1.4
django-cron==0.3.3
django-mathfilters==0.3.0
django-meta==0.2.0
django-scheduler==0.7.2
django-toolbelt==0.0.1
django-widget-tweaks==1.3
elasticsearch==1.1.1
functools32==3.2.3-1
ghdiff==0.4
goose-extractor==1.0.20
gunicorn==19.1.0
hexagonit.swfheader==1.0.1
html2text==2014.7.3
jieba==0.32
jsmin==2.0.11
jsonpickle==0.7.1
lxml==3.3.5
mechanize==0.2.5
nltk==2.0.4
parsedatetime==1.4
poster==0.8.1
psycopg2==2.5.3
python-dateutil==2.2
pytz==2014.4
requests==2.3.0
selenium==2.42.1
setproctitle==1.1.8
simplejson==3.5.3
six==1.7.3
splinter==0.6.0
static3==0.5.1
textblob==0.8.4
tweepy==2.3.0
urllib3==1.9
urlnorm==1.1.2
vobject==0.8.1c
wsgiref==0.1.2
boto==2.19.0
django-storages==1.1.8
django-debug-toolbar==1.2.1
newrelic==2.32.0.28
django-pipeline==1.3.20
whitenoise==1.0.5
html5lib==0.999
python-instagram==1.3.0
